% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Check_Identifiable}
\alias{Check_Identifiable}
\title{Copy of check_estimability}
\usage{
Check_Identifiable(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{\link{check_estimability}} to see list of arguments and usage
}
