% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tmb.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Optimize a TMB model}
\usage{
fit_tmb(obj, fn = obj$fn, gr = obj$gr, startpar = obj$par,
  lower = rep(-Inf, length(startpar)), upper = rep(Inf,
  length(startpar)), getsd = TRUE, control = list(eval.max = 10000,
  iter.max = 10000, trace = 0), bias.correct = FALSE,
  bias.correct.control = list(sd = FALSE, split = NULL, nsplit = NULL,
  vars_to_correct = NULL), savedir = NULL, loopnum = 3,
  newtonsteps = 0, n = Inf, getReportCovariance = FALSE,
  getHessian = FALSE, ...)
}
\arguments{
\item{obj}{The compiled TMB object}

\item{startpar}{Starting values for fixed effects}

\item{lower}{
    vectors of lower and upper bounds, replicated to be as long as
    \code{start}.  If unspecified, all parameters are assumed to be
    unconstrained.
  }

\item{upper}{
    vectors of lower and upper bounds, replicated to be as long as
    \code{start}.  If unspecified, all parameters are assumed to be
    unconstrained.
  }

\item{getsd}{Boolean whether to run standard error calculation}

\item{control}{A list of control parameters. For details see \code{?nlminb}}

\item{bias.correct}{Boolean whether to do epsilon bias-correction}

\item{bias.correct.control}{tagged list of options for epsilon bias-correction, where \code{vars_to_correct} is a character-vector of ADREPORT variables that should be bias-corrected}

\item{savedir}{directory to save results (if \code{savedir=NULL}, then results aren't saved)}

\item{loopnum}{number of times to re-start optimization (where \code{loopnum=3} sometimes achieves a lower final gradient than \code{loopnum=1})}

\item{newtonsteps}{number of extra newton steps to take after optimization (alternative to \code{loopnum})}

\item{n}{sample sizes (if \code{n!=Inf} then \code{n} is used to calculate BIC and AICc)}

\item{getReportCovariance}{Get full covariance matrix of ADREPORTed variables?}

\item{getHessian}{return Hessian for usage in later code}

\item{...}{list of settings to pass to \code{TMB::sdreport}}
}
\value{
the standard output from \code{nlminb}, except with additional diagnostics and timing info, and a new slot containing the output from \code{sdreport}
}
\description{
\code{fit_tmb} runs a TMB model and generates standard diagnostics
}
\examples{
TMBhelper::Optimize( Obj ) # where Obj is a compiled TMB object
}
