% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Optimize.R
\name{Optimize}
\alias{Optimize}
\title{Optimize a TMB model}
\usage{
Optimize(obj, startpar = obj$par, lower = rep(-Inf, length(startpar)),
  upper = rep(Inf, length(startpar)), getsd = TRUE,
  control = list(eval.max = 10000, iter.max = 10000, trace = TRUE), ...)
}
\arguments{
\item{obj}{The compiled TMB object}

\item{startpar}{Starting values for fixed effects}

\item{lower}{lower bounds on fixed effects}

\item{upper}{upper bounds on fixed effects}

\item{getsd}{Boolean whether to run standard error calculation}

\item{control}{list of options to pass to \code{nlminb}}

\item{...}{list of settings to pass to \code{sdreport}}
}
\value{
the standard output from \code{nlminb}, except with additional diagnostics and timing info, and a new slot containing the output from \code{sdreport}
}
\description{
\code{Optimize} runs a TMB model and generates standard diagnostics
}
\examples{
TMBhelper::Optimize( Obj ) # where Obj is a compiled TMB object
}

