% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_re.R
\name{sample_re}
\alias{sample_re}
\title{Sample random effects to correct for re-transformation bias}
\usage{
sample_re(
  obj,
  warmup = 50,
  iter = 150,
  report_names = NULL,
  dat = obj$env$data,
  ...
)
}
\arguments{
\item{obj}{the TMB object after parameter estimation}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations per chain. If step-size adaptation is on (which it is by default), 
    this also controls the number of iterations for which adaptation is run (and
    hence these warmup samples should not be used for inference). The number of 
    warmup iterations should be smaller than \code{iter} and the default is
    \code{iter/2}.}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}

\item{report_names}{which elements of \code{obj$report()} should be recorded;
default \code{report_names=NULL} uses \code{report_names=names(obj$report())}}

\item{...}{adding arguments to pass to \code{\link[tmbstan]{tmbstan}}}
}
\value{
A tagged list containing:
\describe{
  \item{\code{stan_out}}{output from \code{\link[tmbstan]{tmbstan}}}
  \item{\code{report_full}}{A list of output from \code{obj$report()[report_names], except with extra dimension for each MCMC sample}}
  \item{\code{run_time}}{total run time}
}
}
\description{
\code{sample_re} calculates MCMC samples of random effects conditional upon
estimated MLE for fixed effects, and then uses each sample to calculate objects in the report.
This is useful e.g., in correcting for re-transformation bias (by calculating the
posterior mean of a nonlinear transformation of random effects) or visualizing
random-effect variance (which often can be time-consuming using the delta-method
in models with many random effects)
}
\references{
For a discussion of the epsilon-estimator as alternative method
            to correct for re-transformation bias see \url{https://doi.org/10.1016/j.fishres.2015.11.016}
}
