% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBAIC.R
\name{TMBAIC}
\alias{TMBAIC}
\title{Calculate marginal AIC for a fitted model}
\usage{
TMBAIC(opt, k = 2)
}
\arguments{
\item{opt, }{the output from \code{nlminb} or \code{optim}}

\item{k}{the penalty on additional fixed effects (default=2, for AIC)}
}
\value{
AIC, where a parsimonious model has a AIC relative to other candidate models
}
\description{
\code{TMBAIC} calculates AIC for a given model fit
}

