% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Optimize}
\alias{Optimize}
\title{Copy of fit_tmb}
\usage{
Optimize(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?fit_tmb} to see list of arguments and usage
}
